requirements.txt
-i https://pypi.org/simple
colorzero==2.0
gpiozero==1.6.2
i2clcd==0.0.1
libevdev==0.11
paho-mqtt==1.6.1
setuptools==62.6.0
